Update in prep for CI cleanup fix

+++ pytorch_lightning/callbacks/model_checkpoint.py
@@ -16,7 +16,7 @@ import torch
 from pytorch_lightning import _logger as log
 from pytorch_lightning.callbacks.base import Callback
 from pytorch_lightning.utilities import rank_zero_warn, rank_zero_only
-from pytorch_lightning.utilities.cloud_io import gfile, makedirs
+from pytorch_lightning.utilities.cloud_io import gfile, makedirs, is_remote_path
 
 
 class ModelCheckpoint(Callback):
@@ -122,10 +122,10 @@ class ModelCheckpoint(Callback):
             if gfile.isdir(filepath):
                 self.dirpath, self.filename = filepath, '{epoch}'
             else:
-                filepath = os.path.realpath(filepath)
+                if not is_remote_path(filepath):  # dont normalize remote paths
+                    filepath = os.path.realpath(filepath)
                 self.dirpath, self.filename = os.path.split(filepath)
-            if not gfile.exists(self.dirpath):
-                makedirs(self.dirpath)
+            makedirs(self.dirpath)  # calls with exist_ok
         self.save_last = save_last
         self.save_top_k = save_top_k
         self.save_weights_only = save_weights_only
@@ -138,6 +138,7 @@ class ModelCheckpoint(Callback):
         self.best_model_score = 0
         self.best_model_path = ''
         self.save_function = None
+        self.warned_result_obj = False
 
         torch_inf = torch.tensor(np.Inf)
         mode_dict = {
@@ -174,7 +175,12 @@ class ModelCheckpoint(Callback):
                 # dependencies exist then this will work fine.
                 gfile.remove(filepath)
             except AttributeError:
-                os.remove(filepath)
+                if is_remote_path(filepath):
+                    log.warning("Unable to remove stale checkpoints due to running gfile in compatibility mode."
+                                " Please install tensorflow to run gfile in full mode"
+                                " if writing checkpoints to remote locations")
+                else:
+                    os.remove(filepath)
 
     def _save_model(self, filepath, trainer, pl_module):
 
@@ -292,11 +298,26 @@ class ModelCheckpoint(Callback):
         if not gfile.exists(self.dirpath):
             makedirs(self.dirpath)
 
+    def __warn_deprecated_monitor_key(self):
+        using_result_obj = os.environ.get('PL_USING_RESULT_OBJ', None)
+        invalid_key = self.monitor not in ['val_loss', 'checkpoint_on']
+        if using_result_obj and not self.warned_result_obj and invalid_key:
+            self.warned_result_obj = True
+            m = f"""
+                    When using EvalResult(early_stop_on=X) or TrainResult(early_stop_on=X) the
+                    'monitor' key of ModelCheckpoint has no effect.
+                    Remove ModelCheckpoint(monitor='{self.monitor}) to fix')
+                """
+            rank_zero_warn(m)
+
     @rank_zero_only
     def on_validation_end(self, trainer, pl_module):
         # only run on main process
         if trainer.global_rank != 0:
             return
+
+        # TODO: remove when dict results are deprecated
+        self.__warn_deprecated_monitor_key()
 
         metrics = trainer.callback_metrics
         epoch = trainer.current_epoch
